Add first stencil and remove RootNavigating boilerplate
